[#51,markdowndb][s]: test/[...slug].tsx now fetch paths using markdowndb
